Split out treeViewUtils (#116-3)
